requirements.txt
requests==2.27.1
beautifulsoup4==4.12.3
pillow==8.4.0
